# Conflicts:
#	flinkx-connectors/flinkx-connector-jdbc-base/src/main/java/com/dtstack/flinkx/connector/jdbc/source/JdbcInputFormat.java
#	flinkx-core/src/main/java/com/dtstack/flinkx/conf/FlinkxCommonConf.java
#	flinkx-core/src/main/java/com/dtstack/flinkx/conf/SyncConf.java
#	flinkx-core/src/main/java/com/dtstack/flinkx/util/PropertiesUtil.java
#	pom.xml

+++ flinkx-core/src/main/java/com/dtstack/flinkx/outputformat/BaseFileOutputFormat.java
@@ -26,7 +26,7 @@ import com.dtstack.flinkx.conf.BaseFileConf;
 import com.dtstack.flinkx.enums.SizeUnitType;
 import com.dtstack.flinkx.exception.WriteRecordException;
 import com.dtstack.flinkx.sink.WriteMode;
-import org.apache.commons.lang.StringUtils;
+import org.apache.commons.lang3.StringUtils;
 
 import java.io.File;
 import java.io.IOException;
@@ -60,7 +60,7 @@ public abstract class BaseFileOutputFormat extends BaseRichOutputFormat {
     @Override
     public void initializeGlobal(int parallelism) {
         initVariableFields();
-        if(WriteMode.OVERWRITE.name().equalsIgnoreCase(baseFileConf.getWriteMode())){
+        if(WriteMode.OVERWRITE.name().equalsIgnoreCase(baseFileConf.getWriteMode()) && StringUtils.isBlank(baseFileConf.getRestorePath())){
             //Overwrite mode and not delete the data directory first when restoring from a checkpoint
             deleteDataDir();
         }else{
@@ -159,13 +159,13 @@ public abstract class BaseFileOutputFormat extends BaseRichOutputFormat {
     }
 
     @Override
-    protected void commit(long checkpointId) {
+    public void commit(long checkpointId) {
         deleteDataFiles(preCommitFilePathList, tmpPath);
         preCommitFilePathList.clear();
     }
 
     @Override
-    protected void rollback(long checkpointId) {
+    public void rollback(long checkpointId) {
         deleteDataFiles(preCommitFilePathList, outputFilePath);
         preCommitFilePathList.clear();
     }
